Function htmlescape::decode_html_rw

pub fn decode_html_rw<R: BufRead, W: Write>(
    reader: R,
    writer: &mut W
) -> Result<(), DecodeErr>

Decodes an entity-encoded string from a reader to a writer.

Similar to decode_html, except reading from a reader rather than a string, and writing to a writer rather than returning a String.

Arguments

reader - UTF-8 encoded data is read from here.

writer - UTF8- decoded data is written to here.

Errors

Errors can be caused by IO errors, reader producing invalid UTF-8, or by syntax errors.